api/dedicated-storage: import path fixup after module rename
