+++ api/dedicated-storage/client.go
@@ -19,7 +19,7 @@ import (
 	"runtime"
 
 	v1 "github.com/sacloud/sacloud-sdk-go/api/dedicated-storage/apis/v1"
-	"github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	"github.com/sacloud/sacloud-sdk-go/common/saclient"
 )
 
 const (
