+++ api/dedicated-storage/client_test.go
@@ -17,7 +17,7 @@ package dedicatedstorage
 import (
 	"testing"
 
-	"github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	"github.com/sacloud/sacloud-sdk-go/common/saclient"
 )
 
 func TestNewClient(t *testing.T) {
