+++ api/dedicated-storage/contract_test.go
@@ -20,8 +20,8 @@ import (
 	"testing"
 
 	v1 "github.com/sacloud/sacloud-sdk-go/api/dedicated-storage/apis/v1"
+	"github.com/sacloud/sacloud-sdk-go/common/saclient"
 	"github.com/sacloud/sacloud-sdk-go/internal/packages/testutil"
-	"github.com/sacloud/sacloud-sdk-go/internal/saclient"
 )
 
 func TestContract_CRUDL(t *testing.T) {
