+++ api/dedicated-storage/error_test.go
@@ -18,7 +18,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	"github.com/sacloud/sacloud-sdk-go/common/saclient"
 	"github.com/stretchr/testify/require"
 )
