+++ api/dedicated-storage/example_test.go
@@ -21,7 +21,7 @@ import (
 
 	dedicatedstorage "github.com/sacloud/sacloud-sdk-go/api/dedicated-storage"
 	v1 "github.com/sacloud/sacloud-sdk-go/api/dedicated-storage/apis/v1"
-	"github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	"github.com/sacloud/sacloud-sdk-go/common/saclient"
 )
 
 func Example_dedicatedStorageCRUDL() {
